Changed subheader to use title from router
now the subheader is calculated. it is calculated based on the navigationMenu object received from the backend, which contains links and names for links. if this object has a link and a name for the page, then there is no need to specify the title property on the subheader component. the title property is optional