Fix footer covers elements of side menu
The commit fixes an issue where the footer was covering elements of the side menu. The changes include commenting out the left, bottom, and position properties in the .stick-down class in footer.component.scss file. Additionally, in full-layout.component.scss file, the :host::ng-deep.mat-drawer-inner-container selector has been modified to display flex with a column direction and space-between justification.