Add Cypress configuration for TypeScript tests
This commit adds a new file `tsconfig.json` in the `eform-client/cypress` directory. The file includes compiler options for TypeScript tests using Cypress, such as specifying the target as ES5, disabling source maps, and including type definitions for Cypress and Node. Additionally, it specifies to include all files with the extension `.ts`.